requirements.txt
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
coloredlogs==15.0.1
flatbuffers==25.9.23
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-loader==0.4
llvmlite==0.41.1
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
onnxruntime==1.16.3
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.5
pymatting==1.1.14
PyWavelets==1.4.1
referencing==0.35.1
rembg==2.0.61
requests==2.32.4
rpds-py==0.20.1
scikit-image==0.21.0
scipy==1.10.1
sympy==1.13.3
tifffile==2023.7.10
tqdm==4.67.1
urllib3==2.2.3
zipp==3.20.2
